move timestamp under core
