+++ crates/ibc/src/applications/transfer/msgs/transfer.rs
@@ -10,7 +10,7 @@ use ibc_proto::protobuf::Protobuf;
 use crate::applications::transfer::error::TokenTransferError;
 use crate::core::ics04_channel::timeout::TimeoutHeight;
 use crate::core::ics24_host::identifier::{ChannelId, PortId};
-use crate::timestamp::Timestamp;
+use crate::core::timestamp::Timestamp;
 use crate::tx_msg::Msg;
 
 pub const TYPE_URL: &str = "/ibc.applications.transfer.v1.MsgTransfer";
@@ -123,6 +123,8 @@ impl TryFrom<Any> for MsgTransfer {
 
 #[cfg(test)]
 pub mod test_util {
+    use super::*;
+
     use core::ops::Add;
     use core::time::Duration;
     use primitive_types::U256;
@@ -137,7 +139,6 @@ pub mod test_util {
         applications::transfer::BaseCoin,
         core::ics24_host::identifier::{ChannelId, PortId},
         test_utils::get_dummy_bech32_account,
-        timestamp::Timestamp,
     };
 
     // Returns a dummy ICS20 `MsgTransfer`. If no `timeout_timestamp` is
