+++ crates/ibc/src/core/context/chan_close_confirm.rs
@@ -117,10 +117,10 @@ mod tests {
     use crate::core::ics04_channel::Version;
     use crate::core::ics24_host::identifier::{ClientId, ConnectionId};
 
+    use crate::core::timestamp::ZERO_DURATION;
     use crate::mock::client_state::client_type as mock_client_type;
     use crate::mock::context::MockContext;
     use crate::test_utils::DummyTransferModule;
-    use crate::timestamp::ZERO_DURATION;
 
     #[test]
     fn chan_close_confirm_event_height() {
