+++ crates/ibc/src/core/ics03_connection/connection.rs
@@ -18,7 +18,7 @@ use crate::core::ics03_connection::version::Version;
 use crate::core::ics23_commitment::commitment::CommitmentPrefix;
 use crate::core::ics24_host::error::ValidationError;
 use crate::core::ics24_host::identifier::{ClientId, ConnectionId};
-use crate::timestamp::ZERO_DURATION;
+use crate::core::timestamp::ZERO_DURATION;
 
 #[cfg_attr(
     feature = "parity-scale-codec",
