+++ crates/ibc/src/core/ics03_connection/error.rs
@@ -2,7 +2,7 @@ use crate::core::ics02_client::error as client_error;
 use crate::core::ics03_connection::version::Version;
 use crate::core::ics24_host::error::ValidationError;
 use crate::core::ics24_host::identifier::{ClientId, ConnectionId};
-use crate::timestamp::{Timestamp, TimestampOverflowError};
+use crate::core::timestamp::{Timestamp, TimestampOverflowError};
 use crate::Height;
 
 use alloc::string::String;
