+++ crates/ibc/src/core/ics04_channel/handler/chan_close_confirm.rs
@@ -111,9 +111,9 @@ mod tests {
     use crate::core::ics04_channel::Version;
     use crate::core::ics24_host::identifier::{ClientId, ConnectionId};
 
+    use crate::core::timestamp::ZERO_DURATION;
     use crate::mock::client_state::client_type as mock_client_type;
     use crate::mock::context::MockContext;
-    use crate::timestamp::ZERO_DURATION;
 
     use super::validate;
