+++ crates/ibc/src/core/ics04_channel/handler/chan_open_ack.rs
@@ -98,8 +98,8 @@ mod tests {
     use crate::core::ics04_channel::channel::Order;
     use crate::core::ics04_channel::handler::chan_open_ack::validate;
     use crate::core::ics24_host::identifier::ClientId;
+    use crate::core::timestamp::ZERO_DURATION;
     use crate::prelude::*;
-    use crate::timestamp::ZERO_DURATION;
     use rstest::*;
     use test_log::test;
