+++ crates/ibc/src/core/ics04_channel/handler/chan_open_try.rs
@@ -97,9 +97,9 @@ mod tests {
     use crate::core::ics04_channel::msgs::chan_open_try::test_util::get_dummy_raw_msg_chan_open_try;
     use crate::core::ics04_channel::msgs::chan_open_try::MsgChannelOpenTry;
     use crate::core::ics24_host::identifier::{ClientId, ConnectionId};
+    use crate::core::timestamp::ZERO_DURATION;
     use crate::mock::client_state::client_type as mock_client_type;
     use crate::mock::context::MockContext;
-    use crate::timestamp::ZERO_DURATION;
     use crate::Height;
 
     pub struct Fixture {
