+++ crates/ibc/src/core/ics04_channel/handler/recv_packet.rs
@@ -9,7 +9,7 @@ use crate::core::ics24_host::path::{
     AckPath, ChannelEndPath, ClientConsensusStatePath, CommitmentPath, ReceiptPath, SeqRecvPath,
 };
 use crate::core::ics24_host::Path;
-use crate::timestamp::Expiry;
+use crate::core::timestamp::Expiry;
 
 use crate::core::{ContextError, ValidationContext};
 
@@ -183,11 +183,11 @@ mod tests {
     use crate::core::ics04_channel::packet::Packet;
     use crate::core::ics04_channel::Version;
     use crate::core::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
+    use crate::core::timestamp::Timestamp;
+    use crate::core::timestamp::ZERO_DURATION;
     use crate::mock::context::MockContext;
     use crate::mock::ics18_relayer::context::RelayerContext;
     use crate::test_utils::get_dummy_account_id;
-    use crate::timestamp::Timestamp;
-    use crate::timestamp::ZERO_DURATION;
 
     pub struct Fixture {
         pub context: MockContext,
