+++ crates/ibc/src/core/mod.rs
@@ -33,6 +33,7 @@ pub mod ics26_routing;
 pub mod context;
 pub mod events;
 pub mod handler;
+pub mod timestamp;
 
 pub use handler::dispatch;
